extracts effects business logic to a hook
